requirements.txt
requests
pandas
python-dotenv
kafka-python
pyspark
pyarrow
duckdb
apache-airflow